+++ src/main/java/vectorwing/farmersdelight/data/recipe/CraftingRecipes.java
@@ -42,13 +42,13 @@ public class CraftingRecipes
 				.requires(ModTags.CANVAS_SIGNS)
 				.requires(dyeTag)
 				.unlockedBy("has_canvas_sign", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.CANVAS_SIGN.get()))
-				.group("canvas_sign")
+				.group("fd_canvas_sign")
 				.save(consumer);
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.DECORATIONS, hangingCanvasSign, 1)
 				.requires(ModTags.HANGING_CANVAS_SIGNS)
 				.requires(dyeTag)
 				.unlockedBy("has_hanging_canvas_sign", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.HANGING_CANVAS_SIGN.get()))
-				.group("hanging_canvas_sign")
+				.group("fd_hanging_canvas_sign")
 				.save(consumer);
 	}
 
@@ -99,6 +99,7 @@ public class CraftingRecipes
 				.define('e', ForgeTags.EGGS)
 				.define('w', Items.WHEAT)
 				.unlockedBy("has_milk_bottle", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.MILK_BOTTLE.get()))
+				.group("cake")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "cake_from_milk_bottle"));
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.FOOD, Items.CAKE)
 				.requires(ModItems.CAKE_SLICE.get())
@@ -109,6 +110,7 @@ public class CraftingRecipes
 				.requires(ModItems.CAKE_SLICE.get())
 				.requires(ModItems.CAKE_SLICE.get())
 				.unlockedBy("has_cake_slice", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.CAKE_SLICE.get()))
+				.group("cake")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "cake_from_slices"));
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, Items.BOOK)
 				.requires(Items.PAPER)
@@ -227,6 +229,7 @@ public class CraftingRecipes
 				.define('_', Items.OAK_SLAB)
 				.define('D', Items.OAK_TRAPDOOR)
 				.unlockedBy("has_oak_trapdoor", InventoryChangeTrigger.TriggerInstance.hasItems(Items.OAK_TRAPDOOR))
+				.group("fd_cabinet")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModBlocks.BIRCH_CABINET.get())
 				.pattern("___")
@@ -235,6 +238,7 @@ public class CraftingRecipes
 				.define('_', Items.BIRCH_SLAB)
 				.define('D', Items.BIRCH_TRAPDOOR)
 				.unlockedBy("has_birch_trapdoor", InventoryChangeTrigger.TriggerInstance.hasItems(Items.BIRCH_TRAPDOOR))
+				.group("fd_cabinet")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModBlocks.SPRUCE_CABINET.get())
 				.pattern("___")
@@ -243,6 +247,7 @@ public class CraftingRecipes
 				.define('_', Items.SPRUCE_SLAB)
 				.define('D', Items.SPRUCE_TRAPDOOR)
 				.unlockedBy("has_spruce_trapdoor", InventoryChangeTrigger.TriggerInstance.hasItems(Items.SPRUCE_TRAPDOOR))
+				.group("fd_cabinet")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModBlocks.JUNGLE_CABINET.get())
 				.pattern("___")
@@ -251,6 +256,7 @@ public class CraftingRecipes
 				.define('_', Items.JUNGLE_SLAB)
 				.define('D', Items.JUNGLE_TRAPDOOR)
 				.unlockedBy("has_jungle_trapdoor", InventoryChangeTrigger.TriggerInstance.hasItems(Items.JUNGLE_TRAPDOOR))
+				.group("fd_cabinet")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModBlocks.ACACIA_CABINET.get())
 				.pattern("___")
@@ -259,6 +265,7 @@ public class CraftingRecipes
 				.define('_', Items.ACACIA_SLAB)
 				.define('D', Items.ACACIA_TRAPDOOR)
 				.unlockedBy("has_acacia_trapdoor", InventoryChangeTrigger.TriggerInstance.hasItems(Items.ACACIA_TRAPDOOR))
+				.group("fd_cabinet")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModBlocks.DARK_OAK_CABINET.get())
 				.pattern("___")
@@ -267,6 +274,7 @@ public class CraftingRecipes
 				.define('_', Items.DARK_OAK_SLAB)
 				.define('D', Items.DARK_OAK_TRAPDOOR)
 				.unlockedBy("has_dark_oak_trapdoor", InventoryChangeTrigger.TriggerInstance.hasItems(Items.DARK_OAK_TRAPDOOR))
+				.group("fd_cabinet")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModBlocks.MANGROVE_CABINET.get())
 				.pattern("___")
@@ -275,6 +283,7 @@ public class CraftingRecipes
 				.define('_', Items.MANGROVE_SLAB)
 				.define('D', Items.MANGROVE_TRAPDOOR)
 				.unlockedBy("has_mangrove_trapdoor", InventoryChangeTrigger.TriggerInstance.hasItems(Items.MANGROVE_TRAPDOOR))
+				.group("fd_cabinet")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModBlocks.CHERRY_CABINET.get())
 				.pattern("___")
@@ -283,6 +292,7 @@ public class CraftingRecipes
 				.define('_', Items.CHERRY_SLAB)
 				.define('D', Items.CHERRY_TRAPDOOR)
 				.unlockedBy("has_cherry_trapdoor", InventoryChangeTrigger.TriggerInstance.hasItems(Items.CHERRY_TRAPDOOR))
+				.group("fd_cabinet")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModBlocks.BAMBOO_CABINET.get())
 				.pattern("___")
@@ -291,6 +301,7 @@ public class CraftingRecipes
 				.define('_', Items.BAMBOO_SLAB)
 				.define('D', Items.BAMBOO_TRAPDOOR)
 				.unlockedBy("has_bamboo_trapdoor", InventoryChangeTrigger.TriggerInstance.hasItems(Items.BAMBOO_TRAPDOOR))
+				.group("fd_cabinet")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModBlocks.CRIMSON_CABINET.get())
 				.pattern("___")
@@ -299,6 +310,7 @@ public class CraftingRecipes
 				.define('_', Items.CRIMSON_SLAB)
 				.define('D', Items.CRIMSON_TRAPDOOR)
 				.unlockedBy("has_crimson_trapdoor", InventoryChangeTrigger.TriggerInstance.hasItems(Items.CRIMSON_TRAPDOOR))
+				.group("fd_cabinet")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModBlocks.WARPED_CABINET.get())
 				.pattern("___")
@@ -307,12 +319,14 @@ public class CraftingRecipes
 				.define('_', Items.WARPED_SLAB)
 				.define('D', Items.WARPED_TRAPDOOR)
 				.unlockedBy("has_warped_trapdoor", InventoryChangeTrigger.TriggerInstance.hasItems(Items.WARPED_TRAPDOOR))
+				.group("fd_cabinet")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModItems.ROPE.get(), 4)
 				.pattern("s")
 				.pattern("s")
 				.define('s', ModItems.STRAW.get())
 				.unlockedBy("has_straw", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.STRAW.get()))
+				.group("fd_rope")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, ModItems.SAFETY_NET.get(), 1)
 				.pattern("rr")
@@ -323,6 +337,7 @@ public class CraftingRecipes
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.DECORATIONS, ModItems.ROPE.get(), 4)
 				.requires(ModItems.SAFETY_NET.get())
 				.unlockedBy("has_safety_net", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.SAFETY_NET.get()))
+				.group("fd_rope")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "rope_from_safety_net"));
 		ShapedRecipeBuilder.shaped(RecipeCategory.BUILDING_BLOCKS, ModItems.CABBAGE_CRATE.get(), 1)
 				.pattern("###")
@@ -374,6 +389,7 @@ public class CraftingRecipes
 				.requires(ModItems.CANVAS_RUG.get())
 				.requires(ModItems.CANVAS_RUG.get())
 				.unlockedBy("has_canvas_rug", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.CANVAS_RUG.get()))
+				.group("fd_canvas")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "canvas_from_canvas_rug"));
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.BUILDING_BLOCKS, ModItems.ORGANIC_COMPOST.get(), 1)
 				.requires(Items.DIRT)
@@ -387,6 +403,7 @@ public class CraftingRecipes
 				.requires(Items.BONE_MEAL)
 				.unlockedBy("has_rotten_flesh", InventoryChangeTrigger.TriggerInstance.hasItems(Items.ROTTEN_FLESH))
 				.unlockedBy("has_straw", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.STRAW.get()))
+				.group("fd_organic_compost")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "organic_compost_from_rotten_flesh"));
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.BUILDING_BLOCKS, ModItems.ORGANIC_COMPOST.get(), 1)
 				.requires(Items.DIRT)
@@ -400,6 +417,7 @@ public class CraftingRecipes
 				.requires(ModItems.TREE_BARK.get())
 				.unlockedBy("has_tree_bark", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.TREE_BARK.get()))
 				.unlockedBy("has_straw", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.STRAW.get()))
+				.group("fd_organic_compost")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "organic_compost_from_tree_bark"));
 		ShapedRecipeBuilder.shaped(RecipeCategory.BUILDING_BLOCKS, ModItems.TATAMI.get(), 2)
 				.pattern("cs")
@@ -407,12 +425,14 @@ public class CraftingRecipes
 				.define('c', ModItems.CANVAS.get())
 				.define('s', ModItems.STRAW.get())
 				.unlockedBy("has_canvas", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.CANVAS.get()))
+				.group("fd_tatami")
 				.save(consumer);
 
 		// BREAKING DOWN
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.DECORATIONS, ModItems.FULL_TATAMI_MAT.get(), 2)
 				.requires(ModItems.TATAMI.get())
 				.unlockedBy("has_canvas", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.CANVAS.get()))
+				.group("fd_full_tatami_mat")
 				.save(consumer);
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.DECORATIONS, ModItems.HALF_TATAMI_MAT.get(), 2)
 				.requires(ModItems.FULL_TATAMI_MAT.get())
@@ -424,11 +444,13 @@ public class CraftingRecipes
 				.requires(ModItems.HALF_TATAMI_MAT.get())
 				.requires(ModItems.HALF_TATAMI_MAT.get())
 				.unlockedBy("has_canvas", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.CANVAS.get()))
+				.group("fd_full_tatami_mat")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "full_tatami_mat_from_halves"));
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.BUILDING_BLOCKS, ModItems.TATAMI.get(), 1)
 				.requires(ModItems.FULL_TATAMI_MAT.get())
 				.requires(ModItems.FULL_TATAMI_MAT.get())
 				.unlockedBy("has_canvas", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.CANVAS.get()))
+				.group("fd_tatami")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "tatami_block_from_full"));
 	}
 
@@ -472,6 +494,7 @@ public class CraftingRecipes
 				.pattern("##")
 				.define('#', ModItems.STRAW.get())
 				.unlockedBy("has_straw", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.STRAW.get()))
+				.group("fd_canvas")
 				.save(consumer);
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.FOOD, Items.CARROT, 9)
 				.requires(ModItems.CARROT_CRATE.get())
@@ -488,6 +511,7 @@ public class CraftingRecipes
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.FOOD, ModItems.CABBAGE.get(), 9)
 				.requires(ModItems.CABBAGE_CRATE.get())
 				.unlockedBy("has_cabbage_crate", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.CABBAGE_CRATE.get()))
+				.group("fd_cabbage")
 				.save(consumer);
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.FOOD, ModItems.TOMATO.get(), 9)
 				.requires(ModItems.TOMATO_CRATE.get())
@@ -504,6 +528,7 @@ public class CraftingRecipes
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.FOOD, ModItems.RICE.get(), 9)
 				.requires(ModItems.RICE_BAG.get())
 				.unlockedBy("has_rice_bag", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.RICE_BAG.get()))
+				.group("fd_rice")
 				.save(consumer, FarmersDelight.MODID + ":rice_from_bag");
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, ModItems.STRAW.get(), 9)
 				.requires(ModItems.STRAW_BALE.get())
@@ -512,6 +537,7 @@ public class CraftingRecipes
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.FOOD, ModItems.RICE.get())
 				.requires(ModItems.RICE_PANICLE.get())
 				.unlockedBy("has_rice_panicle", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.RICE_PANICLE.get()))
+				.group("fd_rice")
 				.save(consumer);
 	}
 
@@ -543,6 +569,7 @@ public class CraftingRecipes
 				.requires(Items.WHEAT)
 				.requires(Items.WHEAT)
 				.unlockedBy("has_wheat", InventoryChangeTrigger.TriggerInstance.hasItems(Items.WHEAT))
+				.group("fd_wheat_dough")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "wheat_dough_from_water"));
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.FOOD, ModItems.WHEAT_DOUGH.get(), 3)
 				.requires(ForgeTags.EGGS)
@@ -550,6 +577,7 @@ public class CraftingRecipes
 				.requires(Items.WHEAT)
 				.requires(Items.WHEAT)
 				.unlockedBy("has_wheat", InventoryChangeTrigger.TriggerInstance.hasItems(Items.WHEAT))
+				.group("fd_wheat_dough")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "wheat_dough_from_eggs"));
 		ShapedRecipeBuilder.shaped(RecipeCategory.FOOD, ModItems.PIE_CRUST.get(), 1)
 				.pattern("wMw")
@@ -574,6 +602,7 @@ public class CraftingRecipes
 				.requires(ModItems.CABBAGE_LEAF.get())
 				.requires(ModItems.CABBAGE_LEAF.get())
 				.unlockedBy("has_cabbage_leaf", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.CABBAGE_LEAF.get()))
+				.group("fd_cabbage")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "cabbage_from_leaves"));
 		ShapelessRecipeBuilder.shapeless(RecipeCategory.FOOD, ModItems.HORSE_FEED.get(), 1)
 				.requires(Ingredient.of(Items.HAY_BLOCK, ModItems.RICE_BALE.get()))
@@ -613,12 +642,14 @@ public class CraftingRecipes
 				.define('x', Items.SUGAR)
 				.define('O', ModItems.PIE_CRUST.get())
 				.unlockedBy("has_pie_crust", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.PIE_CRUST.get()))
+				.group("fd_apple_pie")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.FOOD, ModItems.APPLE_PIE.get(), 1)
 				.pattern("##")
 				.pattern("##")
 				.define('#', ModItems.APPLE_PIE_SLICE.get())
 				.unlockedBy("has_apple_pie_slice", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.APPLE_PIE_SLICE.get()))
+				.group("fd_apple_pie")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "apple_pie_from_slices"));
 		ShapedRecipeBuilder.shaped(RecipeCategory.FOOD, ModItems.SWEET_BERRY_CHEESECAKE.get(), 1)
 				.pattern("sss")
@@ -628,12 +659,14 @@ public class CraftingRecipes
 				.define('m', ForgeTags.MILK)
 				.define('O', ModItems.PIE_CRUST.get())
 				.unlockedBy("has_pie_crust", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.PIE_CRUST.get()))
+				.group("fd_sweet_berry_cheesecake")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.FOOD, ModItems.SWEET_BERRY_CHEESECAKE.get(), 1)
 				.pattern("##")
 				.pattern("##")
 				.define('#', ModItems.SWEET_BERRY_CHEESECAKE_SLICE.get())
 				.unlockedBy("has_sweet_berry_cheesecake_slice", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.SWEET_BERRY_CHEESECAKE_SLICE.get()))
+				.group("fd_sweet_berry_cheesecake")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "sweet_berry_cheesecake_from_slices"));
 		ShapedRecipeBuilder.shaped(RecipeCategory.FOOD, ModItems.CHOCOLATE_PIE.get(), 1)
 				.pattern("ccc")
@@ -644,12 +677,14 @@ public class CraftingRecipes
 				.define('x', Items.SUGAR)
 				.define('O', ModItems.PIE_CRUST.get())
 				.unlockedBy("has_pie_crust", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.PIE_CRUST.get()))
+				.group("fd_chocolate_pie")
 				.save(consumer);
 		ShapedRecipeBuilder.shaped(RecipeCategory.FOOD, ModItems.CHOCOLATE_PIE.get(), 1)
 				.pattern("##")
 				.pattern("##")
 				.define('#', ModItems.CHOCOLATE_PIE_SLICE.get())
 				.unlockedBy("has_chocolate_pie_slice", InventoryChangeTrigger.TriggerInstance.hasItems(ModItems.CHOCOLATE_PIE_SLICE.get()))
+				.group("fd_chocolate_pie")
 				.save(consumer, new ResourceLocation(FarmersDelight.MODID, "chocolate_pie_from_slices"));
 	}