+++ changelog.md
@@ -6,6 +6,7 @@
 - Knives are now effective at mining Pumpkins and Melons;
 - Knives will harvest **4 Pumpkin Slices** when mining a Pumpkin, instead of a whole unit. Carved Pumpkins are unaffected;
 - Added Canvas Signs to the vanilla `signs` and `hanging_signs` tags;
+- Some crafting recipes are now properly grouped in the Crafting Table's recipe book;
 
 ## 1.2.4